Add Russia grading scale. Make names consistent.
Added the Russia grading scale. Added Japan as an alias for United States. Made some of the scale names more consistent. This will break some old files, so those files will need to be updated to use the new scale names, which should be self-evident from the README. Updated the README.