Add Australia NSW grade scale. Update Australia.
Australia takes letter grades, which seems to be consistent throughout the country. Added the Australia New South Wales grade scale to take number grades specific to NSW.

Applied the same logic to Canada. Canadian variants take number grades specific to the province, but Canada (which is now an alias for United States) is used for letter grades, which also seem to be consistent throughout the country.