Rename India 10 SABCDEFN to India 10 VIT.
This isn't as difficult to type out, and is easier to remember when
you're writing down the scale name. When adding new India 10 scales,
I'll continue to use the university name instead of the letters.

This will break previous files. A simple search-replace from SABCDEFN to
VIT will fix this.