Add numerical 4.0 scale. Modify documentation on China Modified.
As requested, this scale takes in the US 4.0 scale numerically, and
converts all the courses for that student into their overall GPA. It
doesn't verify the GPAs - I can imagine use cases where you'd want to
put numbers besides the typical 0 - 4 +- 0.3.

Also fix an error in the documentation for China Modified, which was the
basis for the creation of this function.