Rename scales overlaid on top of India 10
This is quickly becoming unsustainable, so let's just call them by the
university names like I did for the University of Glasgow. We'll just
continue with this standard from here on out.

This commit breaks previous files. Do a global search replace to replace
`India 10 VIT` to `VIT` and `India 10 BITS` to `BITS Pilani` to fix it.