Begin refactor of India 10 overlay scales
The new file `data/india10.csv` contains data for many universities that
I've encountered. It's infeasible to create a grade scale for every
single one of these universities due to the amount of work.
Additionally, the scales are pretty much the same, so my idea is to
refactor the India 10 code to build a table from the CSV, then attempt
to search through the names of the universities in the table. If there
is a hit, then it selects the correct column in that row based on the
grade input.

This shouldn't take too much work, but I need to also refactor the
initial file reading since this data has commas. I should use a CSV
parsing library instead of my own solution to do this.

I won't be able to work on this until next month, so for now I'm going
to leave the scales I've built until I correctly refactor the code.